Restricting security to a single byte to reduce storage and traffic size.

+++ midspace mod core/data/Scripts/midspace.samplemod/SeModCore/ChatCommandService.cs
@@ -16,7 +16,7 @@
 
         private static readonly Dictionary<string, ChatCommand> Commands = new Dictionary<string, ChatCommand>();
         private static bool _isInitialized;
-        public static uint UserSecurity { get; set; }
+        public static byte UserSecurity { get; set; }
 
         #endregion
